\name{Assignment matrix}
\alias{assignmentMatrix}
\title{Calculate the reads assignment matrix for a gene}
\description{
  TODO
}
\usage{
assignmentMatrix (geneStructure, gene = 1L, readLength, overHang = 1L, 
    paired = FALSE, fragmentProb = NULL, fragmentStart = 0L, 
    normalMean, normalVar, numDevs) 
}
\arguments{
  \item{geneStructure}{}
  \item{gene}{}
  \item{readLength}{}
  \item{overHang}{}
  \item{paired}{}
  \item{fragmentProb}{}
  \item{fragmentStart}{}
  \item{normalMean}{}
  \item{normalVar}{}
  \item{numDevs}{}
}
\details{
  TODO
}
\value{
  TODO
}
% \references
\author{
  Gabor Csardi \email{gcsardi@stat.harvard.edu}
}
\seealso{TODO}
\examples{
## TODO
}
\keyword{misc}

