\name{Splicing complexity}
\alias{geneComplexity}
\title{Complexity of alternative splicing}
\description{
  TODO
}
\usage{
geneComplexity (geneStructure, gene = 1, readLength, overHang = 1L, 
    type = c("relative", "absolute"), norm = c("2", "1", "inf"), 
    paired = FALSE, fragmentProb = NULL, fragmentStart = 0L, 
    normalMean, normalVar, numDevs) 
}
\arguments{
  \item{geneStructure}{TODO}
  \item{gene}{TODO}
  \item{readLength}{TODO}
  \item{overHang}{TODO}
  \item{type}{TODO}
  \item{norm}{TODO}
  \item{paired}{TODO}
  \item{fragmentProb}{TODO}
  \item{fragmentStart}{TODO}
  \item{normalMean}{TODO}
  \item{normalVar}{TODO}
  \item{numDevs}{TODO}
}
\details{
  TODO
}
\value{
  TODO
}
% \references{TODO}
\author{
  Gabor Csardi \email{gcsardi@stat.harvard.edu}
}
% \seealso{TODO}
\examples{
## TODO
}
\keyword{misc}
