\name{matchIso}
\alias{matchIso}
\title{Matching reads to isoforms}
\description{
  TODO
}
\usage{
matchIso (geneStructure, gene = 1, reads, overHang = 1L,
    paired = isPaired(reads), fragmentProb = NULL, fragmentStart = 0L,
    normalMean, normalVar, numDevs) 
}
\arguments{
  \item{geneStructure}{TODO}
  \item{gene}{TODO}
  \item{reads}{TODO}
  \item{overHang}{TODO}
  \item{paired}{TODO}
  \item{fragmentProb}{TODO}
  \item{fragmentStart}{TODO}
  \item{normalMean}{TODO}
  \item{normalVar}{TODO}
  \item{numDevs}{TODO}
}
\details{
  TODO
}
\value{
  TODO
}
% \references{TODO}
\author{
  Gabor Csardi \email{gcsardi@stat.harvard.edu}
}
% \seealso{TODO}
\examples{
## TODO
}
\keyword{misc}
