\name{MISO results}
\alias{MISOresult}
\alias{postMean}
\alias{postMean.MISOresult}
\alias{logLik}
\alias{logLik.MISOresult}
\alias{runData}
\alias{runData.MISOresult}
\title{Results of the MISO algorithm}
\description{
  TODO
}
\usage{
\method{postMean}{MISOresult}(misoResult)
\method{logLik}{MISOresult}(object, \dots)
\method{runData}{MISOresult}(misoResult)
}
\arguments{
  \item{misoResult, object}{TODO}
  \item{\dots}{TODO}
}
\details{
  TODO
}
\value{
  TODO
}
% \references{}
\author{
  Gabor Csardi \email{gcsardi@stat.harvard.edu}
}
% \seealso{}
\examples{
## TODO
}
\keyword{misc}
