\name{Visualization}
\alias{plotIso}
\alias{plotIsoPDF}
\alias{plotIsoSize}
\title{Visualizing gene structures}
\description{
  TODO
}
\usage{
plotIso (geneStructure, gene = 1, xlab = "", ylab = "", \dots)
plotIsoPDF (geneStructure, file, gene = 1, labelwidth = NULL,
     mar = c(0, 0, 2, 0), \dots)
plotIsoSize (geneStructure, gene = 1, labels = NULL,
     stripwidth = 5, stripheight = 1/5) 
}
\arguments{
  \item{geneStructure}{TODO}
  \item{gene}{TODO}
  \item{xlab}{TODO}
  \item{ylab}{TODO}
  \item{file}{TODO}
  \item{labelwidth}{TODO}
  \item{mar}{TODO}
  \item{\dots}{TODO}
  \item{labels}{TODO}
  \item{stripwidth}{TODO}
  \item{stripheight}{TODO}
}
\details{
  TODO
}
\value{
  TODO
}
\references{TODO}
\author{Gabor Csardi \email{gcsardi@stat.harvard.edu}}
\seealso{TODO}
\examples{
## TODO
}
\keyword{misc}
