\name{Read (SAM/BAM) files}
\alias{SAMFile2BAMFile}
\alias{sortBAMFile}
\alias{indexBAMFile}
\title{Manage SAM/BAM files on the disk}
\description{
  TODO
}
\usage{
SAMFile2BAMFile (samfile, bamfile)
sortBAMFile (file, outprefix, key = c("position", "qname")) 
indexBAMFile (file)
}
\arguments{
  \item{samfile}{TODO}
  \item{bamfile}{TODO}
  \item{file}{TODO}
  \item{outprefix}{TODO}
  \item{key}{TODO}
}
\details{
  TODO
}
\value{
  TODO
}
% \references{TODO}
\author{
  Gabor Csardi \email{gcsardi@stat.harvard.edu}
}
\seealso{TODO}
\examples{
## TODO
}
\keyword{misc}
