\name{Number of reads per kilobase}
\alias{RPK}
\title{Calculate the number of reads per kilobase, for a set of genes
  and a set of reads}
\usage{
rpk (genes, reads)
}
\arguments{
  \item{genes}{}
  \item{reads}{}
}
\details{
}
\value{
}
\author{
  Gabor Csardi \email{gcsardi@stat.harvard.edu}
}
% \note{}
% \seealso{}
\examples{
}
\keyword{misc}
