\name{solveIso}
\alias{solveIso}
\title{Simple linear deconvolution for isoform quantitation}
\description{
  TODO
}
\usage{
solveIso (geneStructure, gene = 1L, reads, readLength = getReadLength(reads),
    overHang = 1L, paired = isPaired(reads), fragmentProb = NULL, 
    fragmentStart = 0L, normalMean, normalVar, numDevs) 
}
\arguments{
  \item{geneStructure}{TODO}
  \item{gene}{TODO}
  \item{reads}{TODO}
  \item{readLength}{TODO}
  \item{overHang}{TODO}
  \item{paired}{TODO}
  \item{fragmentProb}{TODO}
  \item{fragmentStart}{TODO}
  \item{normalMean}{TODO}
  \item{normalVar}{TODO}
  \item{numDevs}{TODO}
}
\details{
  TODO
}
\value{
  TODO
}
\references{TODO}
\author{TODO}
\seealso{TODO}
\examples{
## TODO
}
\keyword{misc}
