\name{Assignment matrix}
\alias{assignmentMatrix}
\title{Calculate the reads assignment matrix for a gene}
\description{
  TODO
}
\usage{
assignmentMatrix (geneStructure, gene = 1L, readLength, overHang = 1L, 
    paired = FALSE, fragmentProb = NULL, fragmentStart = 0L, 
    normalMean, normalVar, numDevs) 
}
\arguments{
  \item{geneStructure}{TODO}
  \item{gene}{TODO}
  \item{readLength}{TODO}
  \item{overHang}{TODO}
  \item{paired}{TODO}
  \item{fragmentProb}{TODO}
  \item{fragmentStart}{TODO}
  \item{normalMean}{TODO}
  \item{normalVar}{TODO}
  \item{numDevs}{TODO}
}
\details{
  TODO
}
\value{
  TODO
}
% \references
\author{
  Gabor Csardi \email{gcsardi@stat.harvard.edu}
}
\seealso{TODO}
\examples{
## TODO
}
\keyword{misc}

