\name{Number of reads per kilobase}
\alias{RPK}
\alias{rpk}
\title{Calculate the number of reads per kilobase, for a set of genes
  and a set of reads}
\description{
  TODO
}
\usage{
rpk (genes, reads, overlap=TRUE)
}
\arguments{
  \item{genes}{TODO}
  \item{reads}{TODO}
  \item{overlap}{TODO}
}
\details{
  TODO
}
\value{
  TODO
}
\author{
  Gabor Csardi \email{gcsardi@stat.harvard.edu}
}
% \note{}
% \seealso{}
\examples{
## TODO
}
\keyword{misc}
